requirements.txt
alpaca-trade-api==3.2.0
xgboost==1.7.5
scikit-learn==1.3.0
pandas==2.0.3
yfinance==0.2.28
python-dotenv==1.0.0
numpy==1.24.3
